requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8


langchain-google-genai==2.1.5


google-genai==1.7.0


python-dotenv==1.1.0 
streamlit==1.45.1 
docx2txt==0.9 
faiss-cpu==1.11.0
google-api-python-client==2.170.0 
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
